\name{print.Bayesiangammareg}
\alias{print.Bayesiangammareg}
\title{Print the Bayesian Gamma Regression}
\description{Print the Bayesian Gamma Regression for Joint modeling of Mean 
and Shape}
\usage{
\method{print}{Bayesiangammareg}(x,\ldots)
}
\arguments{
  \item{x}{object of class Bayesiangammareg}
  \item{\dots}{not used.}
}
\value{print the Bayesian Gamma regression}
\references{ 1. Cepeda-Cuervo E. (2001) Modelagem da variabilidade em modelos lineares generalizados. Ph.D. tesis. Instituto de Matematicas. Universidade Federal do Rio do Janeiro.
}

\author{
Arturo Camargo Lozano   \email{bacamargol@unal.edu.co},
Edilberto Cepeda Cuervo \email{ecepedac@unal.edu.co}
}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Mean-Covariance modeling}
\keyword{Gamma}
