\name{print.summary.Bayesiangammareg}
\alias{print.summary.Bayesiangammareg}
\title{Print the Summary of the Bayesian Gamma Regression}
\description{Print the summary Bayesian Gamma regression for Joint modeling of Mean and Shape parameters}
\usage{
\method{print}{summary.Bayesiangammareg}(x, \ldots)
}
\arguments{
  \item{x}{object of class Bayesiangammareg}
  \item{\dots}{not used.}
}
%%\details{}
\value{
Print the summary Bayesian Gamma Regression for Joint modeling of Mean and Shape parameters
}
\references{1. Cepeda-Cuervo E. (2001) Modelagem da variabilidade em modelos lineares generalizados. Ph.D. tesis. Instituto de Matematicas. Universidade Federal do Rio do Janeiro.
}
\author{
Arturo Camargo   \email{bacamargol@unal.edu.co},
Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Mean-Covariance modeling}
\keyword{Gamma}
