\name{summary.Bayesiangammareg}
\alias{summary.Bayesiangammareg}
\title{Print the Bayesian Gamma Regression}
\description{Summarized the Bayesian Gamma Regression for joint modeling
of mean and variance}
\usage{
\method{summary}{Bayesiangammareg}(object, \ldots)
}
\arguments{
  \item{object}{an object of class Bayesiangammareg}
  \item{\dots}{not used.}
}
%%\details{}
\value{
\item{call}{Call}
\item{coefficients}{Coefficients}
\item{deviance}{deviance}
\item{AIC}{AIC}
\item{BIC}{BIC}
}
\references{1. Cepeda-Cuervo E. (2001) Modelagem da variabilidade em modelos lineares generalizados. Ph.D. tesis. Instituto de Matematicas. Universidade Federal do Rio do Janeiro.
2. Cepeda-Cuervo E. and Gamerman D. (2005). Bayesian Methodology for modeling parameters in the two-parameter exponential family. Estadistica 57, 93 105.
3. Cepeda Cuervo E. and Gamerman D. (2001). Bayesian Modeling of Variance Heterogeneity in Normal Regression Models. Brazilian Journal of Probability and Statistics. 14, 207-221.
}
\author{
Brayan Arturo Camargo   \email{bacamargol@unal.edu.co},
Edilberto Cepeda Cuervo \email{ecepedac@unal.edu.co}}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Mean-Covariance modeling}
\keyword{Gamma}
